For merge to Master

+++ ProjektarbeteButik/MainWindow.xaml.cs
@@ -30,7 +30,7 @@ namespace ProjektarbeteButik
         public StackPanel shopInventoryPanel;
         public StackPanel cartInventoryPanel;
         public decimal totalCost;
-        Label cartSubTotalLabel;
+        public Label cartSubTotalLabel;
         public List<Product> productsList = new List<Product>();
         public static Dictionary<Product, int> shoppingCart = new Dictionary<Product, int>();
         public Dictionary<string, string> discountCodes = new Dictionary<string, string>();
@@ -174,16 +174,27 @@ namespace ProjektarbeteButik
             checkOutGrid.ColumnDefinitions.Add(new ColumnDefinition());
             checkOutGrid.ColumnDefinitions.Add(new ColumnDefinition());
 
+            Label couponLabel = new Label
+            {
+                Margin = spacing,
+                Content = "(Optional) Discount Coupon"
+            };
+            checkOutGrid.Children.Add(couponLabel);
+            Grid.SetColumn(couponLabel, 0);
+            Grid.SetRow(couponLabel, 0);
+            Grid.SetColumnSpan(couponLabel, 2);
+
             couponTextBox = new TextBox
             {
                 Margin = spacing,
-                Text = "Optional: Enter Coupon Code here..."
+                TextAlignment = TextAlignment.Center
             };
             checkOutGrid.Children.Add(couponTextBox);
-            Grid.SetColumn(couponTextBox, 0);
+            Grid.SetColumn(couponTextBox, 2);
             Grid.SetRow(couponTextBox, 0);
-            Grid.SetColumnSpan(couponTextBox, 3);
 
+            //We could get rid of this button, and just have the "BUY" button display "Coupon code XXX123 - 25% off, proceed?" (Y/N),
+            //or "Invalid Coupon code, proceed without discount?" (Y/N)
             Button applyDiscountCode = new Button
             {
                 Margin = spacing,
@@ -196,6 +207,7 @@ namespace ProjektarbeteButik
             Grid.SetColumnSpan(applyDiscountCode, 3);
             applyDiscountCode.Click += ApplyDiscountCode_Click;
 
+            //We could also get rid of this button, and just have the cart save automatically whenever it updates
             Button saveCart = new Button
             {
                 Margin = spacing,
@@ -204,7 +216,7 @@ namespace ProjektarbeteButik
             checkOutGrid.Children.Add(saveCart);
             Grid.SetColumn(saveCart, 0);
             Grid.SetRow(saveCart, 1);
-            saveCart.Click += SaveCart_Click;
+            saveCart.Click += SaveCart;
 
             Button checkOutButton = new Button
             {
@@ -218,12 +230,9 @@ namespace ProjektarbeteButik
 
             return checkOutGrid;
         }
-
-        private void ApplyDiscountCode_Click(object sender, RoutedEventArgs e)
-        {
-            
-        }
+        
         public void AddProducts()
+        //We still need to display item description somewhere (included in .csv-file)
         {
             string[] products = File.ReadAllLines("ShopInventory.csv");
             foreach (string s in products)
@@ -256,7 +265,7 @@ namespace ProjektarbeteButik
                 productGrid.Children.Add(addToCart);
                 Grid.SetRow(addToCart, 0);
                 Grid.SetColumn(addToCart, 2);
-                addToCart.Click += AddToCart_Click;
+                addToCart.Click += AddToCart;
 
                 Label productLabel = new Label
                 {
@@ -284,8 +293,7 @@ namespace ProjektarbeteButik
                 Grid.SetColumn(productImage, 0);
             }
         }
-
-        private void AddToCart_Click(object sender, RoutedEventArgs e)
+        private void AddToCart(object sender, RoutedEventArgs e)
         {
             Button button = (Button)sender;
             var product = (Product)button.Tag;
@@ -300,7 +308,6 @@ namespace ProjektarbeteButik
             }
             UpdateCart();
         }
-
         public void UpdateCart()
         {
             totalCost = 0;
@@ -309,7 +316,7 @@ namespace ProjektarbeteButik
             foreach (var item in shoppingCart)
             {
                 totalCost += item.Key.Price * item.Value;
-                cartSubTotalLabel.Content = "Total: $" + totalCost;
+                cartSubTotalLabel.Content = "Subtotal: $" + totalCost;
                 Grid itemGrid = new Grid();
                 itemGrid.RowDefinitions.Add(new RowDefinition { Height = GridLength.Auto });
                 itemGrid.ColumnDefinitions.Add(new ColumnDefinition());
@@ -366,31 +373,20 @@ namespace ProjektarbeteButik
         }
         private void ClearCart(object sender, RoutedEventArgs e)
         {
-            shoppingCart.Clear();
-            File.Delete(CartFilePath);
-            UpdateCart();
-        }
-        private Image CreateImage(string filePath)
-        {
-            ImageSource source = new BitmapImage(new Uri(filePath, UriKind.Relative));
-            Image image = new Image
+            MessageBoxResult result = MessageBox.Show("Clear Shopping Cart, Are You Sure?", "", MessageBoxButton.YesNo);
+            if (result == MessageBoxResult.Yes)
             {
-                Margin = spacing,
-                Source = source,
-                HorizontalAlignment = HorizontalAlignment.Left,
-                VerticalAlignment = VerticalAlignment.Center,
-                MaxHeight = 50,
-                MaxWidth = 50
-            };
-            // A small rendering tweak to ensure maximum visual appeal.
-            RenderOptions.SetBitmapScalingMode(image, BitmapScalingMode.HighQuality);
-            return image;
-        }
+                shoppingCart.Clear();
+                File.Delete(CartFilePath);
+                UpdateCart();    
+            }
+        }        
         public Dictionary<Product, int> LoadCart()
         {
             if (!File.Exists(CartFilePath))
             {
-                MessageBox.Show("No cart to load");
+                //This pop-up gets annoying, do we need it?
+                //MessageBox.Show("No cart to load");
             }
             else
             {
@@ -414,7 +410,7 @@ namespace ProjektarbeteButik
             }
             return shoppingCart;
         }
-        private void SaveCart_Click(object sender, RoutedEventArgs e)
+        private void SaveCart(object sender, RoutedEventArgs e)
         {
             List<string> linesList = new List<string>();
             foreach (KeyValuePair<Product, int> pair in shoppingCart)
@@ -425,6 +421,26 @@ namespace ProjektarbeteButik
             }
             File.WriteAllLines(CartFilePath, linesList);
             MessageBox.Show("Your Cart Was Saved");
+        }
+        private void ApplyDiscountCode_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+        private Image CreateImage(string filePath)
+        {
+            ImageSource source = new BitmapImage(new Uri(filePath, UriKind.Relative));
+            Image image = new Image
+            {
+                Margin = spacing,
+                Source = source,
+                HorizontalAlignment = HorizontalAlignment.Left,
+                VerticalAlignment = VerticalAlignment.Center,
+                MaxHeight = 50,
+                MaxWidth = 50
+            };
+            // A small rendering tweak to ensure maximum visual appeal.
+            RenderOptions.SetBitmapScalingMode(image, BitmapScalingMode.HighQuality);
+            return image;
         }
     }
 }
